meson.build: search for nlohmann_json
yajl is unmaintained and the documented way to include its headers
contradicts with the pkg-config file ("yajl/" prefix or not?).  I'd
like to get rid of yajl, and this is the first step.  nlohmann_json is
an excellent C++ JSON library, let's use that.  This commit only prepares
for using; it contains no code changes.